Remove dependency on libcontainer
