--- Godeps/_workspace/src/github.com/docker/libcontainer/netlink/MAINTAINERS
@@ -1,2 +1,0 @@
-Michael Crosby <michael@crosbymichael.com> (@crosbymichael)
-Guillaume J. Charmes <guillaume@docker.com> (@creack)
