+++ drivers/bridge/netlink_deprecated_linux_armppc64.go
@@ -1,4 +1,6 @@
-package netlink
+// +build arm ppc64
+
+package bridge
 
 func ifrDataByte(b byte) uint8 {
 	return uint8(b)
